Refactored footer components to own files
